Major refactoring - replaced .. paths with @ where possible - skip to content positioning fixed and i18n'ed - component folder structure improved
