+++ src/components/HomeContent.vue
@@ -26,7 +26,7 @@
 
 <script setup lang="ts">
 import {theme} from "@/helpers/theme";
-import SearchBar from "@/components/SearchBar.vue";
+import SearchBar from "@/components/navigation/navbar/SearchBar.vue";
 import {useRouter} from "vue-router";
 
 const router = useRouter();
