+++ src/components/MainPage.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script setup lang='ts'>
-import MainpageCardModal from './MainpageCardModal.vue'
+import MainpageCardModal from '@/components/MainpageCardModal.vue'
 import {useRoute} from "vue-router";
 
 const route = useRoute();
