+++ src/components/navigation/navbar/CourseButton.vue
