+++ src/views/HelpView.vue
@@ -19,10 +19,10 @@
 
 <script lang="ts">
 import {defineComponent} from "vue";
-import Navbar from "../components/navbar/NavbarBase.vue";
-import Help from "../components/HelpCard.vue";
-import Footer from "../components/FooterCard.vue";
-import SkipToContent from "../components/SkipToContent.vue";
+import Navbar from "@/components/navigation/navbar/NavbarBase.vue";
+import Help from "@/components/HelpCard.vue";
+import Footer from "@/components/navigation/FooterCard.vue";
+import SkipToContent from "@/components/helpers/SkipToContent.vue";
 
 export default defineComponent({
   components: {SkipToContent, Navbar, Help, Footer},
