+++ src/views/LeaderboardView.vue
@@ -22,15 +22,9 @@
 }
 </style>
 
-<script lang='ts'>
-import {defineComponent} from "vue";
-import Navbar from "../components/navbar/NavbarBase.vue";
-import Leaderboard from "../components/LeaderboardList.vue";
-import Footer from "@/components/FooterCard.vue";
-import SkipToContent from "@/components/SkipToContent.vue";
-
-export default defineComponent({
-  components: {SkipToContent, Navbar, Leaderboard, Footer },
-  name: "LeaderboardView",
-});
+<script setup lang='ts'>
+import Navbar from "@/components/navigation/navbar/NavbarBase.vue";
+import Leaderboard from "@/components/LeaderboardList.vue";
+import Footer from "@/components/navigation/FooterCard.vue";
+import SkipToContent from "@/components/helpers/SkipToContent.vue";
 </script>
