+++ src/views/MainView.vue
@@ -14,16 +14,9 @@
 <style scoped>
 </style>
 
-<script lang='ts'>
-import {defineComponent} from "vue";
-import Navbar from "../components/navbar/NavbarBase.vue";
-import MainPage from "../components/MainPage.vue";
-import Footer from "@/components/FooterCard.vue";
-import SkipToContent from "@/components/SkipToContent.vue";
-
-
-export default defineComponent({
-  components: {SkipToContent, Footer, Navbar, MainPage},
-  name: "MainView",
-});
+<script lang='ts' setup>
+import Navbar from "@/components/navigation/navbar/NavbarBase.vue";
+import MainPage from "@/components/MainPage.vue";
+import Footer from "@/components/navigation/FooterCard.vue";
+import SkipToContent from "@/components/helpers/SkipToContent.vue";
 </script>
