+++ src/views/NotificationsView.vue
@@ -17,15 +17,9 @@
 }
 </style>
 
-<script lang='ts'>
-import {defineComponent} from "vue";
-import Navbar from "../components/navbar/NavbarBase.vue";
-import Notifications from "../components/NotificationList.vue";
-import Footer from "../components/FooterCard.vue"
-import SkipToContent from "@/components/SkipToContent.vue";
-
-export default defineComponent({
-  components: {SkipToContent, Navbar, Notifications, Footer},
-  name: "NotificationsView",
-});
+<script lang='ts' setup>
+import Navbar from "@/components/navigation/navbar/NavbarBase.vue";
+import Notifications from "@/components/NotificationList.vue";
+import Footer from "@/components/navigation/FooterCard.vue"
+import SkipToContent from "@/components/helpers/SkipToContent.vue";
 </script>
